refactor: Add Blockscount logo svg. Add logo to token card. Move blockscout file to utils
