+++ packages/fether-react/src/Tokens/TokensList/TokenBalance/TokenBalance.js
@@ -4,6 +4,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 import React, { Component } from 'react';
+import { chainName$, isLoading } from '@parity/light.js';
+import light from '@parity/light.js-react';
 import { inject } from 'mobx-react';
 import PropTypes from 'prop-types';
 import { TokenCard } from 'fether-ui';
@@ -11,7 +13,11 @@ import { withRouter } from 'react-router-dom';
 
 import withAccount from '../../../utils/withAccount.js';
 import withBalance from '../../../utils/withBalance';
+import { blockscoutAccountUrl } from '../../../utils/blockscout';
 
+@light({
+  chainName: chainName$
+})
 @withRouter
 @withAccount
 @withBalance
@@ -20,6 +26,19 @@ class TokenBalance extends Component {
   static propTypes = {
     token: PropTypes.object
   };
+
+  openBlockscoutLink () {
+    const { accountAddress, chainName, token } = this.props;
+
+    if (isLoading(chainName) || !accountAddress || !token.address) {
+      return;
+    }
+
+    window.open(
+      blockscoutAccountUrl(accountAddress, chainName, token.address),
+      '_blank'
+    );
+  }
 
   handleClick = () => {
     const { accountAddress, history, sendStore, token } = this.props;
@@ -31,7 +50,13 @@ class TokenBalance extends Component {
   };
 
   render () {
-    return <TokenCard onClick={this.handleClick} {...this.props} />;
+    return (
+      <TokenCard
+        openBlockscoutLink={() => this.openBlockscoutLink()}
+        onClick={this.handleClick}
+        {...this.props}
+      />
+    );
   }
 }
