+++ packages/fether-react/src/assets/sass/components/_token.scss
@@ -36,6 +36,15 @@
     height: ms(0) * 1.3;
   }
 
+  .token_blockscout_logo {
+    margin-left: 0.7rem;
+    img {
+      display: block;
+      width: 2rem;
+      height: auto;
+    }
+  }
+
   .token_symbol {
     height: ms(0) * 1.3;
     font-size: ms(-2);
